+++ omnisend/includes/SDK/V1/class-contact.php
@@ -37,9 +37,9 @@ class Contact {
 	private $phone_consent = null;
 
 	private $email_opt_in_source     = null;
-	private $email_status            = null;
+	private $email_status            = 'nonSubscribed';
 	private $phone_opt_in_source     = null;
-	private $phone_status            = null;
+	private $phone_status            = 'nonSubscribed';
 	private array $custom_properties = array();
 
 	/**
@@ -129,23 +129,13 @@ class Contact {
 			'tags'        => array_values( array_unique( $this->tags ) ),
 		);
 
-		$email_consent_status = $this->email_opt_in_source ? 'subscribed' : 'nonSubscribed';
-		if ( $this->email_status == 'unsubscribed' ) {
-			$email_consent_status = 'unsubscribed';
-		}
-
-		$sms_consent_status = $this->phone_opt_in_source ? 'subscribed' : 'nonSubscribed';
-		if ( $this->phone_status == 'unsubscribed' ) {
-			$sms_consent_status = 'unsubscribed';
-		}
-
 		if ( $this->email ) {
 			$email_identifier = array(
 				'type'     => 'email',
 				'id'       => $this->email,
 				'channels' => array(
 					'email' => array(
-						'status'     => $email_consent_status,
+						'status'     => $this->email_status,
 						'statusDate' => $time_now,
 					),
 				),
@@ -172,7 +162,7 @@ class Contact {
 				'id'       => $this->phone,
 				'channels' => array(
 					'sms' => array(
-						'status'     => $sms_consent_status,
+						'status'     => $this->phone_status,
 						'statusDate' => $time_now,
 					),
 				),
@@ -198,10 +188,6 @@ class Contact {
 
 		if ( $this->last_name ) {
 			$arr['lastName'] = $this->last_name;
-		}
-
-		if ( $this->phone ) {
-			$arr['phone'] = $this->phone;
 		}
 
 		if ( $this->address ) {
@@ -374,6 +360,15 @@ class Contact {
 	}
 
 	/**
+	 * Gets contact email.
+	 *
+	 * @return string
+	 */
+	public function get_email(): string {
+		return $this->email;
+	}
+
+	/**
 	 * Sets contact id.
 	 *
 	 * @param $id
@@ -398,6 +393,15 @@ class Contact {
 	}
 
 	/**
+	 * Gets contact gender.
+	 *
+	 * @return string
+	 */
+	public function get_gender(): string {
+		return $this->gender;
+	}
+
+	/**
 	 * Sets contact first_name.
 	 *
 	 * @param $first_name
@@ -409,6 +413,15 @@ class Contact {
 	}
 
 	/**
+	 * Gets contact first_name.
+	 *
+	 * @return string
+	 */
+	public function get_first_name(): string {
+		return $this->first_name;
+	}
+
+	/**
 	 * Sets contact last_name.
 	 *
 	 * @param $last_name
@@ -417,6 +430,15 @@ class Contact {
 	 */
 	public function set_last_name( $last_name ): void {
 		$this->last_name = $last_name;
+	}
+
+	/**
+	 * Gets contact last_name.
+	 *
+	 * @return string
+	 */
+	public function get_last_name(): string {
+		return $this->last_name;
 	}
 
 	/**
@@ -433,6 +455,15 @@ class Contact {
 	}
 
 	/**
+	 * Gets contact address.
+	 *
+	 * @return string
+	 */
+	public function get_address(): string {
+		return $this->address;
+	}
+
+	/**
 	 * Sets contact city.
 	 *
 	 *
@@ -442,6 +473,15 @@ class Contact {
 	 */
 	public function set_city( $city ): void {
 		$this->city = $city;
+	}
+
+	/**
+	 * Gets contact city.
+	 *
+	 * @return string
+	 */
+	public function get_city(): string {
+		return $this->city;
 	}
 
 	/**
@@ -457,6 +497,15 @@ class Contact {
 	}
 
 	/**
+	 * Gets contact state.
+	 *
+	 * @return string
+	 */
+	public function get_state(): string {
+		return $this->state;
+	}
+
+	/**
 	 * Sets contact country.
 	 *
 	 *
@@ -466,6 +515,15 @@ class Contact {
 	 */
 	public function set_country( $country ): void {
 		$this->country = $country;
+	}
+
+	/**
+	 * Gets contact country.
+	 *
+	 * @return string
+	 */
+	public function get_country(): string {
+		return $this->country;
 	}
 
 	/**
@@ -480,6 +538,15 @@ class Contact {
 	}
 
 	/**
+	 * Gets contact postal_code.
+	 *
+	 * @return string
+	 */
+	public function get_postal_code(): string {
+		return $this->postal_code;
+	}
+
+	/**
 	 * Sets contact set_phone.
 	 *
 	 * @param $phone
@@ -487,6 +554,15 @@ class Contact {
 	 */
 	public function set_phone( $phone ): void {
 		$this->phone = $phone;
+	}
+
+	/**
+	 * Gets contact phone.
+	 *
+	 * @return string
+	 */
+	public function get_phone(): string {
+		return $this->phone;
 	}
 
 	/**
@@ -498,6 +574,15 @@ class Contact {
 	 */
 	public function set_birthday( $birthday ): void {
 		$this->birthday = $birthday;
+	}
+
+	/**
+	 * Gets contact birthday.
+	 *
+	 * @return string
+	 */
+	public function get_birthday(): string {
+		return $this->birthday;
 	}
 
 	/**
@@ -513,7 +598,7 @@ class Contact {
 	}
 
 	/**
-	 * Sets email opt in source. It's used to track where contact opted in to receive emails. It's required to mark contact email as subscribed.
+	 * Sets email opt in source and email status. It's used to track where contact opted in to receive emails. It's required to mark contact email as subscribed.
 	 *
 	 * Common format is `form:form_name` or `popup:popup_name`.
 	 *
@@ -523,24 +608,38 @@ class Contact {
 	 */
 	public function set_email_opt_in( $opt_in_text ): void {
 		$this->email_opt_in_source = $opt_in_text;
+		$this->email_status        = 'subscribed';
 	}
 
 	/**
-	 * Sets email status. It's used to track whether contact opted in to receive emails.
-	 *
-	 * Allowed values are subscribed, nonsubscribed and unsubscribed
-	 *
-	 * @param string $email_status
+	 * Sets email status as subscribed.
 	 *
 	 * @return void
 	 */
-	public function set_email_status( string $email_status ): void {
-		$this->email_status = $email_status;
+	public function set_email_subscriber(): void {
+		$this->email_status = 'subscribed';
 	}
 
+	/**
+	 * Sets email status as unsubscribed.
+	 *
+	 * @return void
+	 */
+	public function set_email_unsubscriber(): void {
+		$this->email_status = 'unsubscribed';
+	}
 
 	/**
-	 * Sets phone opt in source. It's used to track where contact opted in to receive emails. It's required to mark contact phone as subscribed.
+	 * Gets contact email_status.
+	 *
+	 * @return string
+	 */
+	public function get_email_status(): string {
+		return $this->email_status;
+	}
+
+	/**
+	 * Sets phone opt in source and phone status. It's used to track where contact opted in to receive emails. It's required to mark contact phone as subscribed.
 	 *
 	 * Common format is `form:form_name` or `popup:popup_name`.
 	 *
@@ -550,19 +649,34 @@ class Contact {
 	 */
 	public function set_phone_opt_in( $opt_in_text ): void {
 		$this->phone_opt_in_source = $opt_in_text;
+		$this->phone_status        = 'subscribed';
 	}
 
 	/**
-	 * Sets phone status. It's used to track whether contact opted in to receive sms.
-	 *
-	 * Allowed values are subscribed, nonsubscribed and unsubscribed
-	 *
-	 * @param string $phone_status
+	 * Sets phone status as subscribed.
 	 *
 	 * @return void
 	 */
-	public function set_phone_status( string $phone_status ): void {
-		$this->phone_status = $phone_status;
+	public function set_phone_subscriber(): void {
+		$this->phone_status = 'subscribed';
+	}
+
+	/**
+	 * Sets phone status as unsubscribed.
+	 *
+	 * @return void
+	 */
+	public function set_phone_unsubscriber(): void {
+		$this->phone_status = 'unsubscribed';
+	}
+
+	/**
+	 * Gets contact phone_status.
+	 *
+	 * @return string
+	 */
+	public function get_phone_status(): string {
+		return $this->phone_status;
 	}
 
 	/**
@@ -591,7 +705,6 @@ class Contact {
 		$this->phone_consent = $consent_text;
 	}
 
-
 	/**
 	 * @param $key
 	 * @param $value
@@ -612,6 +725,15 @@ class Contact {
 	}
 
 	/**
+	 * Gets contact custom_properties.
+	 *
+	 * @return array
+	 */
+	public function get_custom_properties(): array {
+		return $this->custom_properties;
+	}
+
+	/**
 	 * @param $tag
 	 * @param bool $clean_up_tag clean up tag to be compatible with Omnisend
 	 *
@@ -627,5 +749,14 @@ class Contact {
 		}
 
 		$this->tags[] = $tag;
+	}
+
+	/**
+	 * Gets contact tags.
+	 *
+	 * @return array
+	 */
+	public function get_tags(): array {
+		return $this->tags;
 	}
 }
